-- CEF-376 Added condition to the "Certify and Submit to SLT" button on the Report Summary screen. Enables the button only when the validation status is PASSED or PASSED_WARNINGS, and only when user is certifier.
Manual refresh is needed to update the validation status in the summary page after QA, otherwise the status will remain the same. This may be addressed with CEF-375.

A new ticket should be created for the validation status to be unvalidated after any edits are made. Currently edits can be made after QA is executed and validation status remains the same, allowing users to submit unvalidated data.